Lagde oppsummeringssteg
* Gamle index.tsx renames til [[...step]].tsx
* Bruker får opp oppsummeringsside hvis de går på
  /utfylling/oppsummering
* Spørsmålene er flyttet inn i en egen Utfyllingssteg-komponent
